requirements-test.txt
# Test dependencies — install with: pip install -r requirements-test.txt
pytest>=7.4
pytest-mock>=3.11
requests-mock>=1.11
pytest-asyncio>=0.23
python-dotenv>=1.0
requests>=2.31
